requirements.txt
# Prediction Market Trading System - Python Dependencies
# Versions pinned from prediction-trader-env (pip freeze Feb 2026)

# Core Dependencies
numpy>=2.4.2
pandas>=3.0.0
scipy>=1.17.0

# HTTP & API
requests>=2.32.5
aiohttp>=3.13.3
websockets>=16.0

# Mathematical Optimization
cvxpy>=1.8.1
ortools>=9.15.0

# Data Storage
sqlalchemy>=2.0.46
redis>=7.1.1

# Configuration
pyyaml>=6.0.3
python-dotenv>=1.2.1

# Logging & Monitoring
loguru>=0.7.3

# Web Dashboard
flask>=3.1.2

# WSL browser launcher (sudo apt install wslu)
# wslu provides wslview command

# Time & Date
pytz>=2025.2
python-dateutil>=2.9.0

# Testing & Development
pytest>=9.0.2
pytest-asyncio>=1.3.0

# Visualization
matplotlib>=3.10.8
plotly>=6.5.2

# Machine Learning (optional)
scikit-learn>=1.8.0

# Utilities
tqdm>=4.67.3
click>=8.3.1
